new project bug fix
